MarkerObject: name should be test, not bytes
Ran into this while migrating some code using MarkerObjects to Python 3. The
name of the MarkerObject is supposed to be a human-readable label, so it
should be text, not bytes.

This change enforces that the name is str (i.e., text) in Python 3. In Python 2,
we allow both but issue a warning when using str (i.e., bytes in Python 3).

Passes tests in Python 2.7 and 3.5 on OS X.