qcore.inspection: Remove cache from get_full_name (#86)
The cache is broken for classes: if we set the cache on a base class
and then try to get the full name of a child class, it will use the
base's value. We could fix this by checking only __dict__, but it seems
better to just avoid the cache. Callers should add caching if performance
is critical.